+++ graphbrain/tests/test_logic.py
@@ -9,6 +9,8 @@ class TestLogic(unittest.TestCase):
         self.hg.destroy()
         edge = hedge('(is/Pd.sc (the/M sun/C) red/C)')
         self.hg.add(edge)
+        edge = hedge('(is/Pd.sc red/C color/C)')
+        self.hg.add(edge)
 
     def tearDown(self):
         self.hg.close()
@@ -30,6 +32,18 @@ class TestLogic(unittest.TestCase):
     def test_eval_rule(self):
         rule = hedge("""
         (:- (prop/P ENTITY PROP) (is/Pd.sc ENTITY PROP))
-        """.strip())
-        result = list(match[0] for match in eval_rule(self.hg, rule))
-        self.assertEqual(result, [hedge('(prop/P (the/M sun/C) red/C)')])
+        """)
+        result = list(inference.edge for inference in eval_rule(self.hg, rule))
+        self.assertEqual(
+            result, [hedge('(prop/P (the/M sun/C) red/C)'),
+                     hedge('(prop/P red/C color/C)')])
+
+    def test_eval_rule_and(self):
+        rule = hedge("""
+        (:- (color/P ENTITY COLOR)
+            (and (is/Pd.sc (the/M ENTITY) COLOR)
+                 (is/Pd.sc COLOR color/C)))
+        """)
+        result = list(inference.edge for inference in eval_rule(self.hg, rule))
+        self.assertEqual(
+            result, [hedge('(color/P sun/C red/C)')])