+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 - Argument role pattern matching.
 - HG.match().
 - Hyperedge.atom().
-- Logic engine.
+- Logic engine, HG.eval().
 
 ### Changed
 - parse() now returns a dictionary that includes inferred edges.